# Transfer strings in a string table