--- a/src/devtools/store.js
+++ b/src/devtools/store.js
@@ -630,6 +630,22 @@ export default class Store extends EventEmitter {
     // We'll use the parent ID to adjust selection if it gets deleted.
 
     let i = 2;
+
+    // Reassemble the string table.
+    let stringTable = [
+      null, // ID = 0 corresponds to the null string.
+    ];
+    const stringTableSize = operations[i++];
+    const stringTableEnd = i + stringTableSize;
+    while (i < stringTableEnd) {
+      let nextLength = operations[i++];
+      let nextString = utfDecodeString(
+        (operations.slice(i, i + nextLength): any)
+      );
+      stringTable.push(nextString);
+      i += nextLength;
+    }
+
     while (i < operations.length) {
       const operation = operations[i];
       switch (operation) {
@@ -686,23 +702,13 @@ export default class Store extends EventEmitter {
             ownerID = ((operations[i]: any): number);
             i++;
 
-            const displayNameLength = operations[i];
+            const displayNameStringID = operations[i];
+            const displayName = stringTable[displayNameStringID];
             i++;
-            const displayName =
-              displayNameLength === 0
-                ? null
-                : utfDecodeString(
-                    (operations.slice(i, i + displayNameLength): any)
-                  );
-            i += displayNameLength;
 
-            const keyLength = operations[i];
+            const keyStringID = operations[i];
+            const key = stringTable[keyStringID];
             i++;
-            const key =
-              keyLength === 0
-                ? null
-                : utfDecodeString((operations.slice(i, i + keyLength): any));
-            i += +keyLength;
 
             if (__DEBUG__) {
               debug(