--- a/src/devtools/store.js
+++ b/src/devtools/store.js
@@ -632,14 +632,14 @@ export default class Store extends EventEmitter {
     let i = 2;
 
     // Reassemble the string table.
-    let stringTable = [
+    const stringTable = [
       null, // ID = 0 corresponds to the null string.
     ];
     const stringTableSize = operations[i++];
     const stringTableEnd = i + stringTableSize;
     while (i < stringTableEnd) {
-      let nextLength = operations[i++];
-      let nextString = utfDecodeString(
+      const nextLength = operations[i++];
+      const nextString = utfDecodeString(
         (operations.slice(i, i + nextLength): any)
       );
       stringTable.push(nextString);