When a large tree mounts or updates, in most cases strings are repetitive. For example, when interacting with a real complex app, I'm seeing how in every batch of `operations`, only ~10% of displayNames/keys are unique. The other ~90% are the same names that we encode multiple times. In the past, we've optimized encoding them to be cached. However, decoding is still lazy. It turns out that **decoding repetitive strings accounts for 40% of time spend in Store's `operations` handler**.

This fixes it with a string table. Individual events like `ADD` no longer contain the encoded strings. Instead, each batch of `operations` contains a list of unique strings in the beginning, and then further operations mention them by ID. I updated the `OVERVIEW` to explain the new format.

## Synthetic results

**Before:** 40ms in renderer, 61ms in the store
**After:** 30ms in renderer, 34ms in the store

## Product code results

**Before:** 109ms in renderer, 246ms in the store
**After:** 108ms in renderer, 150ms in the store